Fixed code review [22110]
